+++ packages/vite-plugin/src/codegen/shared.ts
@@ -1,31 +1,40 @@
 // SPDX-FileCopyrightText: con terra GmbH and contributors
 // SPDX-License-Identifier: Apache-2.0
-import { dirname } from "node:path";
+import { dirname, basename } from "node:path/posix";
+import { normalizePath } from "vite";
 
+const APP_MODULE = "@@open-pioneer-app";
+
+const APP_META_QUERY = "open-pioneer-app";
 const APP_META_RE = /[?&]open-pioneer-app(?:$|&)/;
+
+const APP_PACKAGES_QUERY = "open-pioneer-packages";
 const APP_PACKAGES_RE = /[?&]open-pioneer-packages(?:$|&)/;
+
+const APP_CSS_QUERY = "open-pioneer-styles";
 const APP_CSS_RE = /[?&]open-pioneer-styles(?:$|&)/;
+
+const APP_I18N_INDEX_QUERY = "open-pioneer-i18n-index";
 const APP_I18N_INDEX_RE = /[?&]open-pioneer-i18n-index(?:$|&)/;
+
+const APP_I18N_QUERY = "open-pioneer-i18n";
 const APP_I18N_RE = /[?&]open-pioneer-i18n(?:$|&)/;
 const APP_I18N_LOCALE_RE = /[?&]locale=(?<locale>.*?)(?:$|&)/;
-const APP_HOOKS_RE = /[/\\]@@open-pioneer-react-hooks(?:\?|$)/;
+
+const PACKAGE_HOOKS_MODULE = "@@open-pioneer-react-hooks";
+
 const SOURCE_FILE_RE = /^(.*?)(?:\?|$)/;
 
-export const APP_META_QUERY = "open-pioneer-app";
-export const APP_PACKAGES_QUERY = "open-pioneer-packages";
-export const APP_CSS_QUERY = "open-pioneer-styles";
-export const APP_I18N_INDEX_QUERY = "open-pioneer-i18n-index";
-export const APP_I18N_QUERY = "open-pioneer-i18n";
-export const PACKAGE_HOOKS = "@@open-pioneer-react-hooks";
+export type VirtualModule = VirtualAppModule | VirtualI18nMessages | VirtualPackageModule;
 
 export interface VirtualAppModule {
     type: "app-meta" | "app-packages" | "app-css" | "app-i18n-index";
-    importer: string;
+    packageDirectory: string;
 }
 
 export interface VirtualI18nMessages {
     type: "app-i18n";
-    importer: string;
+    packageDirectory: string;
     locale: string;
 }
 
@@ -34,52 +43,85 @@ export interface VirtualPackageModule {
     packageDirectory: string;
 }
 
-export function parseVirtualAppModuleId(
-    moduleId: string
-): VirtualAppModule | VirtualI18nMessages | VirtualPackageModule | undefined {
-    if (moduleId.match(APP_HOOKS_RE)) {
-        return {
-            type: "package-hooks",
-            packageDirectory: dirname(getSourceFile(moduleId))
-        };
+/**
+ * Takes a module id as input and parses it.
+ * If this plugin is not responsible to load the module id, this function returns `undefined`.
+ *
+ * Otherwise, the module is parsed into an object and returned.
+ */
+export function parseVirtualModuleId(inputModuleId: string): VirtualModule | undefined {
+    const moduleId = normalizePath(inputModuleId);
+    if (!moduleId || moduleId.includes("\\0")) {
+        return undefined;
     }
 
-    if (moduleId.match(APP_META_RE)) {
-        const type = "app-meta";
-        const importer = getSourceFile(moduleId);
-        return { type, importer };
-    } else if (moduleId.match(APP_PACKAGES_RE)) {
-        const type = "app-packages";
-        const importer = getSourceFile(moduleId);
-        return { type, importer };
-    } else if (moduleId.match(APP_CSS_RE)) {
-        const type = "app-css";
-        const importer = getSourceFile(moduleId);
-        return { type, importer };
-    } else if (moduleId.match(APP_I18N_INDEX_RE)) {
-        const type = "app-i18n-index";
-        const importer = getSourceFile(moduleId);
-        return { type, importer };
-    } else if (moduleId.match(APP_I18N_RE)) {
-        const type = "app-i18n";
-        const importer = getSourceFile(moduleId);
-        const locale = moduleId.match(APP_I18N_LOCALE_RE)?.groups?.["locale"];
-        if (!locale) {
-            throw new Error("Missing locale name in i18n module request");
-        }
-        return { type, importer, locale };
+    const sourceFile = getSourceFile(moduleId); // module id with out query
+    if (!sourceFile) {
+        return undefined;
+    }
+
+    const base = basename(sourceFile);
+    if (base === PACKAGE_HOOKS_MODULE) {
+        return {
+            type: "package-hooks",
+            packageDirectory: dirname(sourceFile)
+        };
+    }
+    if (base === APP_MODULE) {
+        return parseAppModuleId(sourceFile, moduleId);
     }
     return undefined;
 }
 
-export function appLocaleFileId(appModuleId: string, locale: string) {
-    return `${appModuleId}?${APP_I18N_QUERY}&locale=${locale}`;
+function parseAppModuleId(sourceFile: string, moduleId: string): VirtualModule | undefined {
+    const packageDirectory = dirname(sourceFile);
+    if (moduleId.match(APP_META_RE)) {
+        const type = "app-meta";
+        return { type, packageDirectory };
+    } else if (moduleId.match(APP_PACKAGES_RE)) {
+        const type = "app-packages";
+        return { type, packageDirectory };
+    } else if (moduleId.match(APP_CSS_RE)) {
+        const type = "app-css";
+        return { type, packageDirectory };
+    } else if (moduleId.match(APP_I18N_INDEX_RE)) {
+        const type = "app-i18n-index";
+        return { type, packageDirectory };
+    } else if (moduleId.match(APP_I18N_RE)) {
+        const type = "app-i18n";
+        const locale = moduleId.match(APP_I18N_LOCALE_RE)?.groups?.["locale"];
+        if (!locale) {
+            throw new Error("Missing locale name in i18n module request");
+        }
+        return { type, packageDirectory, locale };
+    }
+    throw new Error(`Unexpected module id: ${moduleId}`);
+}
+
+/**
+ * Serializes the given virtual module descriptor into a module id string.
+ */
+export function serializeModuleId(mod: VirtualModule): string {
+    switch (mod.type) {
+        case "package-hooks":
+            return `${mod.packageDirectory}/${PACKAGE_HOOKS_MODULE}`;
+        case "app-meta":
+            return `${mod.packageDirectory}/${APP_MODULE}?${APP_META_QUERY}`;
+        case "app-packages":
+            return `${mod.packageDirectory}/${APP_MODULE}?${APP_PACKAGES_QUERY}`;
+        case "app-css":
+            return `${mod.packageDirectory}/${APP_MODULE}?${APP_CSS_QUERY}`;
+        case "app-i18n-index":
+            return `${mod.packageDirectory}/${APP_MODULE}?${APP_I18N_INDEX_QUERY}`;
+        case "app-i18n":
+            return `${mod.packageDirectory}/${APP_MODULE}?${APP_I18N_QUERY}&locale=${mod.locale}`;
+    }
 }
 
 function getSourceFile(moduleId: string) {
     const sourceFile = moduleId.match(SOURCE_FILE_RE)?.[1];
     if (!sourceFile || moduleId[0] == "\0") {
-        throw new Error(`Failed to get actual source file from '${moduleId}'.`);
+        return undefined;
     }
     return sourceFile;
 }