+++ packages/vite-plugin/test-data/codegen-css/packages/style1/build.config.mjs
@@ -1,3 +1,3 @@
 export default {
-    styles: "./styles.css"
+    styles: "./styles.scss"
 }